break hub-content code into modules
