+++ packages/content/src/index.ts
@@ -2,4 +2,5 @@
  * Apache-2.0 */
 
 export * from "./content";
-export * from "./util";
+export * from "./hub";
+export * from "./portal";
