--- packages/content/src/util.ts
@@ -1,11 +1,0 @@
-/**
- * Parse item ID and layer ID (if any) from dataset record ID
- *
- * @param datasetId Hub API dataset record id ({itemId}_{layerId} or {itemId})
- */
-export function parseDatasetId(
-  datasetId: string
-): { itemId: string; layerId?: string } {
-  const [itemId, layerId] = datasetId ? datasetId.split("_") : [];
-  return { itemId, layerId };
-}
